requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
dotenv==0.9.9
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.28.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
multidict==6.6.3
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parsel==1.10.0
pillow==11.3.0
playwright==1.54.0
propcache==0.3.2
Protego==0.5.0
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==13.0.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
queuelib==1.8.0
referencing==0.36.2
requests==2.32.4
requests-file==2.1.0
rpds-py==0.26.0
rsa==4.9.1
Scrapy==2.13.3
seaborn==0.13.2
service-identity==24.2.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.47.1
tenacity==8.5.0
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zope.interface==7.2
